backend/requirements.txt
# Kipawa Platform - Backend Dependencies
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and Geospatial
sqlalchemy[asyncio]==1.4.52
geoalchemy2==0.14.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Configuration
pyyaml==6.0.1
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0

# Geospatial Processing
shapely==2.0.2
fiona==1.9.5
rasterio==1.3.9
pyproj==3.6.1
geopandas==0.14.1

# Satellite Data
sentinelhub==3.9.0
requests==2.31.0
aiohttp==3.9.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# EUDR Compliance
defusedxml==0.7.1
cryptography==41.0.7

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Logging
numpy<2.0
loguru==0.7.2
structlog==23.2.0
